take jekyll-grid theme for TOH front matter
Turns out two jekyll instances will be used, this second one sets the
front page matter of the Open Hyre board, and relatedly the layout of
newsletter online issues (for now) and of task description pages. It's
a jekyll-grid instance and should be up at www.openhyre.org/issue. The
previous theme adopted, Wall-E, will be kept for blog posts and for a
blog home grid mimicking the job board, to run at blog.openhyre.org.

This project is becoming concrete for me and I finally have a grounded
confidence in the many ways it can go right. Already I'm getting hearty
replies in the emails and it's only the fourth response. Setting up all
the little front-facing things isn't feeling like chore but a game of
presentation, it's helping and I can obsess over texts and names for a
bit.

Today is very bright with a mild climate.