change style to make TOH front page more compact
The whitespace on the board home was pushing tasks below the fold and
generally bothering. I mustered up changes in widths and seems to be
cool for smaller screens, I'll know for sure when it's on GH Pages.

There's a drive to work on this all the time and the benefit of short
vacations. I skipped the voice acting course today, again. Yesterday
my friend sent my referral for his new remote gig, it'd be at least
some comfort to be on the way with the selection process, which he
said was simple, in case reality hits sooner than expected or The Open
Hyre doesn't garner much attention.

This same friend, producer and coder, now I realize, is a great person
to ask for feedback on this project. Everything is suddenly so smooth
and easier, falling into place. The nights are harder lately but I'm on
the run now. There's something booming in my chest again.