gear up for what it means to be on the front
I'm getting doses of self-doubt along with the ecstasy.

The code for tagging multiple tracks came along with a puff and it was
a nice feeling, as was saving the first session in easy format (1. 2.
3.) which together make for commit material soon.

I'm getting urges again, after a long interday period of extreme calm,
about hustling this efficiently in bulk. For this matter I'll duplicate
functionality or even code from the currently evolving `play.py` program
into a `soci.py` program that can manage lists of people contacts and
act as repo to save messages and also for canned message generation both
with clipboard integration and finally a workflow step-by-step not unlike
the flow ClojureScript library which will be ported to Hy for hyre at
some point.

Also been thinking about a `pass.py` that can generate ssh and gpg keys
(using `plumbum` delegating to shell) and help with pass and clipboard.
This program can then integrate with `soci.py` to secure sensitive files.

I'll gear up on the social mgmt interface and start selling this week, if
only to face real-world responses and play out these anxieties.